Removal of many dual LaTeX and latex2html section titles in the manual.
As the HTML manual (http://www.nmr-relax.com/manual/index.html) is now compiled with Unicode
support, the Greek characters in the titles are now supported.  Therefore in the model-free and the
values, gradients, and Hessians chapters, the dual LaTeX and latex2html section titles could be
collapsed to the standard LaTeX section title.  This will result in better formatting of the manual
and its links.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@26238 b7916896-f9f9-0310-9fe5-b3996d8957d5