Fix for the poorly written User_functions.test_structure_add_atom GUI test.
This fixes one part of 2 of the bug #22772 (https://gna.org/bugs/?22772), the modelfree4 binary
issue and the User_functions GUI tests with wxPython 2.9 failures of the test suite.

The problem was that a list element was being set in the GUI test, but that element did not exist
yet.  Somehow this worked in wxPython 2.8.  But the bad code failed on wxPython 2.9.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@26254 b7916896-f9f9-0310-9fe5-b3996d8957d5