Fix for the find next bug in the relax controller window.
This is bug #22815 (https://gna.org/bugs/?22815), the failure of find next using F3 (or Ctrl-G on
Mac OS X) in the relax controller window if search text has already been set.

The fix was simple, as the required flags are in the self.find_data class object (an instance of
wx.FindReplaceData).



git-svn-id: svn://svn.gna.org/svn/relax/trunk@26291 b7916896-f9f9-0310-9fe5-b3996d8957d5