Fix for find dialog in the relax controller window.
This is for bug #22816 (https://gna.org/bugs/?22816), the find functionality of the relax controller
window does not find text when using wxPython >= 2.9.

The find wxPython events are now bound to the find dialog rather than the relax controller window
LogCtrl element for displaying the relax messages.  This works on all wxPython versions.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@26292 b7916896-f9f9-0310-9fe5-b3996d8957d5