Improvement for all wizards and user functions in the relax GUI.
The focus is now set on the currently displayed page of the wizard.  This allows the keyboard to be
active without requiring a mouse click.  Now text can be instantly input into the first text control
and the tab key can jump between elements.  As the GUI user functions are wizards with a single
page, this is a significant usability improvement for the GUI.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@26295 b7916896-f9f9-0310-9fe5-b3996d8957d5