Modified the internal structural object backend for the structure.read_pdb user function.
The MolContainer._det_pdb_element() method for handling PDB files with missing element information
has been updated to use the Periodic_table.has_element() method to check if the PDB atom name
corresponds to any atoms in the periodic table.  This allows for far greater support for HETATOMS
and all of the metals.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@26324 b7916896-f9f9-0310-9fe5-b3996d8957d5