Fix for bug #22938 (https://gna.org/bugs/?22938).
This is the failure of the test suite in the relax GUI.  The problem was that the
status.skip_blacklisted_tests variable did not exist - it was only initialised if relax is started
in test suite mode.  Now the value is always set from within the status module and defaults to True.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@26495 b7916896-f9f9-0310-9fe5-b3996d8957d5