Manual Python 3 fixes for the dict.key() function which returns a list or iterator in Python 2 or 3.
This involves a number of changes.  The biggest is the conversion of the "x in y.keys()" statements
to "x in y".  For code which requires a list of keys, the function calls "list(y.keys())" or
preferably "sorted(y.keys())" are used throughout (sorted() ensures that the list will be of the
same order on all operating systems and Python implementations).  A number of "x in list(y.keys())"
statements were simplified to "x in y", some list() calls changed to sorted(), and some unnecessary
list() calls were removed.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@26519 b7916896-f9f9-0310-9fe5-b3996d8957d5