Important bug fix for racing causing the GUI to freeze.
This is really only seen in the GUI tests on MS Windows systems, as a user could never be fast
enough with the mouse.  The GUI interpreter flush() method for ensuring that all user functions in
the queue have been cleared now calls wx.Yield() to force all wxPython events to also be flushed.
This change will avoid random freezing of the relax test suite.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@26543 b7916896-f9f9-0310-9fe5-b3996d8957d5