Fix for bug #22961 (https://gna.org/bugs/?22961)
This is the failure of relaxation data loading with the message "IndexError: list index out of
range".  The bug was found by Julien Orts.

It is triggered by loading relaxation data from a file containing spin name information and
supplying the spin ID using the spin name to restrict data loading to a spin subset.  To solve the
problem, the pipe_control.relax_data.pack_data() function has been redesigned.  Now the selection
union concept of Chris MacRaild's selection object is being used by joining the spin ID constructed
from the data file and the user supplied spin ID with '&', and using this to isolate the correct
spin system.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@26588 b7916896-f9f9-0310-9fe5-b3996d8957d5