Big Python 3 bug fix for the dep_check module for the detection of the NMRPipe showApod software.
The showApod program was falsely detected as always not being present when using Python 3.  This is
because the output of the program was being tested using string comparisons.  However the output
from programs obtained from the subprocess module is no longer strings but rather byte-arrays in
Python 3.  Therefore the byte-array is not being converted to text if Python 3 is being used,
allowing the showApod software to be detected.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@26667 b7916896-f9f9-0310-9fe5-b3996d8957d5