Modified the printouts from the structure.write_pdb user function if models are present.
Instead of printing out 'MODEL', 'ATOM, HETATM, TER' and 'ENDMDL' for each model, the header 'MODEL
records' is printed followed by a single '.' character for each model.  For structures with many models,
this results in a huge speed up of the user function which is strongly limited by how fast the
terminal can display text.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@26782 b7916896-f9f9-0310-9fe5-b3996d8957d5