Bug fix for the pymol.view user function for when no PDB file exists.
The pymol.view user function would fail with an AttributeError when the currently loaded data does
not exist as a PDB file.  This is now caught and the non-existent PDB is no longer displayed.  A
better solution might be to dump all the current structural data into a temporary file and load
that, all within a try-finally statement to be sure to delete the temporary file.  This solution may
not be what the user is interested in anyway.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@26798 b7916896-f9f9-0310-9fe5-b3996d8957d5