Cleared the list of blacklisted objects for the cdp.exp_info data structure.
The data_store.exp_info.ExpInfo class blacklist variable had previously not been used.  But after
recent changes, the list was now active.  As all the contents of the container were blacklisted, the
container was being initialised as being empty when reading the XML formatted state or results
files.  Therefore the blacklist is now set to an empty list.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@27040 b7916896-f9f9-0310-9fe5-b3996d8957d5