Shifted the atomic assembly code from the structure.align user function into its own function.
The new function assemble_coordinates() of the pipe_control.structure.main module will be used to
standardise the process of assembling atomic coordinates for all of the structure user functions.
This will improve the support for comparing different molecules rather than different models as
missing atoms or divergent primary sequence are properly handled, and it has multi-pipe support.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@27067 b7916896-f9f9-0310-9fe5-b3996d8957d5