Changed the design of the lib.structure.internal.coordinates.assemble_coord_array() function.
The elements_flag argument has been renamed to seq_info_flag.  If this is set, then in addition to
the atomic elements, the molecule name, residue name, residue number, and atom name is now assembled
and returned.  This information is now the common information between the structures, hence the
return values for the elements are a list of str rather than list of lists.

All of the code in pipe_control.structure.main has been updated for the change.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@27091 b7916896-f9f9-0310-9fe5-b3996d8957d5