Modified some system tests of the structure.align and structure.superimpose user functions.
The displace_id argument has been introduced for both of these user functions for finer control over
which atoms are translated and rotated by the algorithm.  This allows, for example, to align
structures based on a set of backbone heavy atoms while the protons and side chains are displaced by
default.  Or if a domain is aligned, then just that domain can be displaced.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@27100 b7916896-f9f9-0310-9fe5-b3996d8957d5