Significantly expanded the gnuplot script from via the correlation_matrix() plotting API function.
This is for the structure.atomic_fluctuations user function.  The output terminal is now set to EPS,
the colour map changed from the default to a blue-red map, labels have been added, the plot is now
square, and comments are now included throughout the script to help a user hand modify it after
creation.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@27134 b7916896-f9f9-0310-9fe5-b3996d8957d5