Fix for bug #23187 (https://gna.org/bugs/?23187).
This is the problem whereby opening the spin viewer window, deleting a residue, and then reopening
the spin viewer crashes relax.

This change completes the spin viewer update_*() functions.  The prune_list variable was initialised
but not used.  Now it is used to store the keys of the items to delete, and then the items are
deleted at the end in a new loop so that the loop over the dictionary keys is not corrupted.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@27167 b7916896-f9f9-0310-9fe5-b3996d8957d5