More robustness for the spin viewer GUI window prune_*() methods.
The data returned from the self.tree.GetItemPyData(key) call can in rare racing cases not contain
the 'id' key.  This is now being checked for and are being skipped in the prune_mol(), prune_res()
and prune_spin() methods.

The problem was found in the Mf.test_bug_20479_gui_final_pipe system test when running the command:
for i in {1..10}; do ./relax --gui-tests --time -d &>> gui_tests.log; done



git-svn-id: svn://svn.gna.org/svn/relax/trunk@27192 b7916896-f9f9-0310-9fe5-b3996d8957d5