Bug fix for the printouts from the relax_data.read user function.
This problem was introduced in the last relax release (at r26588).  The problem is that the spin ID
in the loaded relaxation data printout is the same for all data, being the spin ID of the first
spin.  This has no effect on how relax runs, it is only incorrect feedback.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@27296 b7916896-f9f9-0310-9fe5-b3996d8957d5