Bug fix for the PDB secondary sheet handling when combining multiple PDB structures.
The helix and sheet metadata now converts the original chain IDs into molecule indices, shifted to
new values based on the currently loaded data, when the structure.read_pdb user function is
executed.  When the structure.write_pdb user function is executed, the molecule indices are
converted into new chain IDs.  This allows the chain IDs in the HELIX and SHEET records to match
those of the ATOM and HETATOM records.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@27322 b7916896-f9f9-0310-9fe5-b3996d8957d5