Large simplification of the atomic coordinate assembly code in the internal structural object.
This is in the lib.structure.internal.coordinates.assemble_coord_array() function.  The logic of the
function has recently changed due to the introduction of the pairwise sequence alignments.  This
caused a lot of code to now be redundant, and also incorrect in certain cases.  This simplification
fixes the problem caught by the Structure.test_align_molecules2 system test.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@27338 b7916896-f9f9-0310-9fe5-b3996d8957d5